backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
pymysql==1.1.0
sqlalchemy==2.0.41
pandas==2.3.1
python-dotenv==1.1.1